experiment parsing v-1
see #TODO in write_run_template
