+++ omnigan/norms.py
@@ -1,3 +1,5 @@
+"""Normalization layers used in blocks
+"""
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
