+++ shared/experiment/exp_defaults.yaml
@@ -1,0 +1,24 @@
+experiment:
+  exp_desc: Training d only - 1 # Description of the experiment
+  defaults: shared/trainer/defaults.yaml # default file to configure trainers
+  config: config/trainer/local_tests.yaml # file to override defaults trainers
+  no_comet: false # don't log on comet?
+  dev_mode: false # just trying?
+  base_dir: /network/tmp1/$USER/omnigan/experiments/ # experiment base dir in which runs will be stored
+  repeat: 1 # if > 1, repeat the list in runs (to sample differently from the same conf for instance)
+
+runs:
+  - sbatch: # dictionary for the sbatch job
+      # --------------------
+      # -----  SBATCH  -----
+      # --------------------
+      cpu: 6 # how many cpus
+      mem: 32 # how much RAM
+      gpu: ":1" # what kind of GPU (--gres=gpu<run.gpu>)
+      duration: "12:00:00" # quotes are needed otherwise parsed as time in seconds not string
+      partition: unkillable # partition for the sbatch job
+    trainer: # exploration dictionary (runs.trainer > config_file > defaults.yaml)
+      # ---------------------
+      # -----  TRAINER  -----
+      # ---------------------
+      tasks: [d]
