+++ tests/test_trainer.py
@@ -4,7 +4,8 @@ from pathlib import Path
 
 sys.path.append(str(Path(__file__).parent.parent.resolve()))
 from omnigan.trainer import Trainer
-from omnigan.utils import freeze, load_test_opts
+from omnigan.utils import load_test_opts
+from omnigan.tutils import freeze
 from run import print_header
 
 parser = argparse.ArgumentParser()
